[#2] Docs : main page 컴포넌트 분리
