--- src/components/main/MainCategory.tsx
@@ -1,5 +1,0 @@
-function MainCategory() {
-	return <div>MainCategory</div>;
-}
-
-export default MainCategory;
