--- src/components/main/MainDataList.tsx
@@ -1,5 +1,0 @@
-function MainDataList() {
-	return <div>MainDataList</div>;
-}
-
-export default MainDataList;
