--- src/components/main/MainModal.tsx
@@ -1,5 +1,0 @@
-function MainModal() {
-	return <div>MainModal</div>;
-}
-
-export default MainModal;
