+++ src/types/type.d.ts
@@ -1,0 +1,15 @@
+export type DataType = {
+	idx: number;
+	name: string;
+	mainImage: string;
+	description: string;
+	spaceCategory: string;
+	price: number;
+	maximumPurchases: number;
+	registrationDate: string;
+};
+
+export type FilteringType = {
+	price: string;
+	region: string;
+};
